pyproject.toml
[project]
name = "embedding-collapse-study"
version = "0.1.0"
description = "Empirical study of embedding space geometry in contrastive learning"
requires-python = ">=3.10"

dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "lightly>=1.5.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.0",
    "umap-learn>=0.5.6",
    "wandb>=0.16.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "tqdm>=4.66.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "ruff>=0.4.0",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 88
target-version = "py310"
